fixes win32 alignment problems

+++ compiler/options.nim
@@ -720,7 +720,7 @@ proc `$`*(c: IdeCmd): string =
   of ideKnown: "known"
   of ideMsg: "msg"
 
-proc floatInt64Align*(conf: ConfigRef): int =
+proc floatInt64Align*(conf: ConfigRef): int16 =
   ## Returns either 4 or 8 depending on reasons.
   if conf.target.targetCPU == cpuI386:
     #on Linux/BSD i386, double are aligned to 4bytes (except with -malign-double)
